moved all direct firebase calls out of home component
