+++ src/Home.js
@@ -3,8 +3,7 @@ import { Text, TouchableWithoutFeedback, TextInput, View, FlatList, Button, Safe
 import Platforms from './Platforms';
 import styles from './Home.screen.style';
 import { sortAlphabetically, sortNumerically } from './Utilities';
-import {clientID, bearer} from './config.js';
-import { fetchPlats, fetchGames } from './firebaseConnect';
+import * as Api from './crud.js';
 
 export default function HomeScreen({navigation}){
     const [games, setGames] = useState(null);
@@ -13,111 +12,43 @@ export default function HomeScreen({navigation}){
     const [platform, setPlatform] = useState(null);
     const [isLoading, setIsLoading] = useState(true);
     const [searchTerm, setSearchTerm] = useState('');
+    
     let searchText = '';
     const limit = 20;
 
-    
-    /*
-    const fetchGames = async (searchTerm, platformID) => {
-
-      const whereSearch = platformID ? `where platforms = (${platformID});` : '';
-
-      // clean searchTerm here
-      const text = `fields name,
-        alternative_names.name,
-        platforms.name,
-        version_parent,
-        release_dates.game.name,
-        release_dates.platform.name,
-        release_dates.region,
-        release_dates.date,
-        version_title;
-        limit ${limit};
-        search "${searchTerm}";
-        ${whereSearch};`
-
-      const request = new Request(`https://api.igdb.com/v4/games`, {
-        method: "POST",
-        headers:{
-          "Client-ID": clientID,
-          "Authorization":`Bearer ${bearer}`,
-        },
-        body:text
-      });
-
+    const getGames = async (searchTerm, platformID) => {
       try{
         setIsLoading(true);
-        
-        const apiCall = await fetch(request);
-        const games = await apiCall.json();
+        const games = await Api.getGames(searchTerm, platformID)
         setIsLoading(false);
-  
-        // save to useState
         setGames(games);
-  
-      } catch (error){
-        console.log(error);
+      } catch(error){
+          console.error(error);
+          setIsLoading(false);
       }
-    }
-    */
 
-    /*
-    const fetchPlatforms = async () => {
-      const text = `fields name, generation; where category= (1,5,6); limit 200;`;
-      const request = new Request(`https://api.igdb.com/v4/platforms`, {
-        method: "POST",
-        headers:{
-          "Accept":"application/json",
-          "Client-ID": clientID,
-          "Authorization":`Bearer ${bearer}`,
-        },
-        body:text
-      });
+    };
 
+    const getPlatforms = async () => {
       try {
-        const apiCall = await fetch(request);
-        const retrievedPlatforms = await apiCall.json();
-        console.log(retrievedPlatforms);
-
-        // filter out all non-generation platforms before sorting numerically, otherwise the sort will not work
-        //const filterGeneration = retrievedPlatforms.filter((item => 'generation' in item)).sort((a,b) => b.generation - a.generation);
-        setAvailablePlatforms(retrievedPlatforms.sort(sortAlphabetically('name')));
-
-      } catch(error) {
-
+        const platforms = await Api.getPlatforms('');
+        const sortNameAscending = sortAlphabetically('name');
+        setAvailablePlatforms(platforms.sort(sortNameAscending));
+      } catch (error) {
+        console.error(error);
       }
-    }
-    */
-
-    const fetchGames = async (searchTerm, platformID) => {
-     try{
-      setIsLoading(true);
-      const games = await fetchGames({platformID: '', searchTerm: 'Skyrim'})
-      setIsLoading(false);
-      setGames(games);
-    } catch(error){
-        console.log(error);
-    }
-
-    }
-
-    const fetchPlatforms = async () => {
-      fetchPlats('')
-      .then(result => setAvailablePlatforms(result.data.sort(sortAlphabetically('name'))))
-      .catch(error => console.log(error));
-    }
+    };
 
     useEffect(() => {
-      // get available platforms just the first time the component is loaded
-      fetchPlatforms();
+      getPlatforms();
     }, []);
     
     useEffect(() => {
-      fetchGames(searchTerm, platform);
-      // changing to include platform select
+      // new search needed when platform filter is changed, or a new search term is used
+      getGames(searchTerm, platform);
     },[searchTerm, platform]);
 
-    // returns comma separated list of platforms from an arry - move to utility function
+    // returns comma separated list of platforms from an array - move to utility function
     const platforms = (platforms) => {
       if (!platforms) {
         return '';
@@ -127,7 +58,6 @@ export default function HomeScreen({navigation}){
     };
   
     const renderItem = ({item}) => ( 
-
       <TouchableWithoutFeedback style={styles.item} onPress={() => navigation.navigate("Details", {id: item.id.toString()})}>
         <View style={styles.item}>
           <Text style={styles.itemID}>{item.id}</Text>
@@ -135,7 +65,6 @@ export default function HomeScreen({navigation}){
           <Text style={styles.platforms}>{platforms(item.platforms)}</Text>
         </View>
        </TouchableWithoutFeedback>
-      
     );
 
     const onChangeText = (text) => {
